requirements.txt
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dotenv==0.21.1
tabulate==0.9.0
pytest==7.2.2
attrs==22.2.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
